+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 1.1 - 2/19/2025
 - Variables for project title and version
+- Added menus
+- Removed Exit button
+- Changed initalization
 
 # 1.0 - 2/19/2025
 - Project compiling