+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Changed initalization
 - Added Saving Build flags to a text file
 - Added Loading Build flags to a text file
+- Added Trace logs to a text file
 
 # 1.0 - 2/19/2025
 - Project compiling